+++ api/helpers/things.js
@@ -7,6 +7,73 @@ const moment = require("moment");
 const { okToFlipFeatured } = require("./user");
 const { as, db, sql } = require("./db");
 
+// Define the keys we're testing (move these to helper/things.js ?
+const titleKeys = ["id", "title"];
+const shortKeys = titleKeys.concat([
+  "type",
+  "images",
+  "post_date",
+  "updated_date"
+]);
+const mediumKeys = shortKeys.concat(["body", "bookmarked", "location"]);
+const thingKeys = mediumKeys.concat([
+  "original_language",
+  "published",
+  "files",
+  "videos",
+  "featured",
+  "tags",
+  "url"
+]);
+const caseKeys = thingKeys.concat([
+  "issue",
+  "communication_mode",
+  "communication_with_audience",
+  "content_country",
+  "decision_method",
+  "end_date",
+  "facetoface_online_or_both",
+  "facilitated",
+  "voting",
+  "number_of_meeting_days",
+  "ongoing",
+  "start_date",
+  "total_number_of_participants",
+  "targeted_participant_demographic",
+  "kind_of_influence",
+  "targeted_participants_public_role",
+  "targeted_audience",
+  "participant_selection",
+  "specific_topic",
+  "staff_type",
+  "type_of_funding_entity",
+  "typical_implementing_entity",
+  "typical_sponsoring_entity",
+  "who_else_supported_the_initiative",
+  "who_was_primarily_responsible_for_organizing_the_initiative"
+]);
+const methodKeys = thingKeys.concat([
+  "best_for",
+  "communication_mode",
+  "decision_mode",
+  "facilitated",
+  "governance_contribution",
+  "issue_interdependency",
+  "issue_polarization",
+  "issue_technical_complexity",
+  "kind_of_influence",
+  "method_of_interaction",
+  "public_interaction_method",
+  "typical_funding_source",
+  "typical_implementing_entity",
+  "typical_sponsoring_entity"
+]);
+const organizationKeys = thingKeys.concat([
+  "executive_director",
+  "issue",
+  "sector"
+]);
+
 function addRelatedList(owner_type, owner_id, related_type, id_list) {
   // TODO: escape id_list to avoid injection attacks
   if (!id_list || !id_list.length) {
@@ -230,13 +297,12 @@ function getEditXById(type) {
               key: as.name(key),
               value: as.videos(newThing[key])
             });
-          } else if (key === "lead_image") {
-            var img = newThing[key];
+          } else if (key === "images") {
             updatedThingFields.push({
               key: as.name(key),
-              value: as.attachment(img)
+              value: as.strings(newThing[key])
             });
-          } else if (["other_images", "files"].includes(key)) {
+          } else if (key === "files") {
             updatedThingFields.push({
               key: as.name(key),
               value: as.attachments(newThing[key])
@@ -331,5 +397,12 @@ module.exports = {
   diffRelatedList,
   difference,
   getEditXById,
-  supportedTypes
+  supportedTypes,
+  titleKeys,
+  shortKeys,
+  mediumKeys,
+  thingKeys,
+  caseKeys,
+  methodKeys,
+  organizationKeys
 };
