+++ api/sql/featuredmap.sql
@@ -10,7 +10,7 @@ SELECT
   featured,
   title,
   to_json(COALESCE(location, '("","","","","","","","","")'::geolocation)) AS location,
-  to_json(COALESCE(lead_image, '("","",0)'::attachment)) AS lead_image
+  to_json(COALESCE(images, '{}')) AS images
 FROM things, localized_texts
 WHERE things.id = localized_texts.thingid AND
       things.hidden = false AND
