+++ api/sql/list_map_cases.sql
@@ -2,7 +2,7 @@ SELECT
     cases.id,
     'case' as type,
     to_json(cases.location) AS location,
-    to_json(cases.lead_image) AS lead_image,
+    to_json(cases.images) AS images,
     localized_texts.title AS title
 FROM
     cases,
